+++ backend/bruno/GET -repertoire-{id}.yml
@@ -5,7 +5,7 @@ info:
 
 http:
   method: GET
-  url: http://localhost:8080/repertoire/deeb0eb6-1e9d-4490-a73b-486cea58cd78
+  url: http://localhost:8080/repertoire/deeb0eb6-1e9d-4490-a73b-486cea58cd70
   auth: inherit
 
 settings:
