+++ backend/main.go
@@ -3,90 +3,35 @@ package main
 import "net/http"
 import "encoding/json"
 import "log"
-import "os"
 import "database/sql"
-import "github.com/go-sql-driver/mysql"
-import "github.com/google/uuid"
-import "github.com/joho/godotenv"
-
-func connectDb() *sql.DB {
-	log.Println("fetching config...")
-
-	dbuser := os.Getenv("DBUSER")
-	dbpass := os.Getenv("DBPASS")
-
-	cfg := mysql.NewConfig()
-	cfg.User = dbuser
-	cfg.Passwd = dbpass
-	cfg.Net = "unix"
-	cfg.Addr = "/tmp/mysql.sock"
-	cfg.DBName = "chessrepeat"
-
-	log.Println("setting up connection...")
-
-	var db *sql.DB
-	var err error
-	db, err = sql.Open("mysql", cfg.FormatDSN())
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	log.Println("testing connection...")
-
-	pingErr := db.Ping()
-	if pingErr != nil {
-		log.Fatal(pingErr)
-	}
-
-	log.Println("connected to database!")
-
-	return db
-}
-
-func isValidName(name string) bool {
-	return name != ""
-}
-
-func isValidTrainAs(trainAs string) bool {
-	return trainAs == "white" || trainAs == "black"
-}
 
 func main() {
-
-	err := godotenv.Load()
-	if err != nil {
-		log.Fatal("Error loading .env file")
-	}
-
 	log.Println("starting server...")
 
 	var db = connectDb()
 
 	http.HandleFunc("/repertoire/{id}", func(w http.ResponseWriter, r *http.Request) {
 		if r.Method == "GET" {
-			unparsedId := r.PathValue("id")
-			var id, err = uuid.Parse(unparsedId)
+			var id, err = parseIdFromRequest(r)
 			if err != nil {
+				log.Println("invalid repertoire:", err)
 				w.WriteHeader(http.StatusBadRequest)
 				return
 			}
 
-			type repertoireJson struct {
-				RepertoireId *string `json:"id"`
-				Name         *string `json:"name"`
-				TrainAs      *string `json:"trainAs"`
-			}
-
 			log.Println("fetching repertoire for id:", id)
 
-			row := db.QueryRow("SELECT * FROM repertoire WHERE repertoire_id=?", id)
-
-			var repertoire repertoireJson
-			err = row.Scan(&repertoire.RepertoireId, &repertoire.Name, &repertoire.TrainAs)
-			if err != nil {
-				// TODO(ben): make not found 404s
-				w.WriteHeader(http.StatusInternalServerError)
-				return
+			repertoire, err := fetchRepertoire(db, id);
+			if (err != nil) {
+				if err == sql.ErrNoRows {
+					log.Println("no repertoire found for id:", id, err)
+					w.WriteHeader(http.StatusNotFound)
+					return
+				} else {
+					log.Println("database error when fetching id:", id, err)
+					w.WriteHeader(http.StatusInternalServerError)
+					return
+				}
 			}
 
 			w.Header().Set("Content-Type", "application/json")
@@ -94,62 +39,24 @@ func main() {
 
 			w.WriteHeader(http.StatusOK)
 		} else if r.Method == "POST" {
-			unparsedId := r.PathValue("id")
-			var id, err = uuid.Parse(unparsedId)
+			var repertoire, err = parseRepertoireFromRequest(r)
 			if err != nil {
+				log.Println("invalid repertoire:", err)
 				w.WriteHeader(http.StatusBadRequest)
 				return
 			}
 
-			type repertoireJson struct {
-				Name    *string `json:"name"`
-				TrainAs *string `json:"trainAs"`
-			}
+			log.Println("creating repertoire:", repertoire)
 
-			decoder := json.NewDecoder(r.Body)
-			var repertoire repertoireJson
-			err = decoder.Decode(&repertoire)
+			repertoire, err = createRepertoire(db, repertoire)
 			if err != nil {
-				w.WriteHeader(http.StatusBadRequest)
-				return
-			}
-			if !isValidName(*repertoire.Name) {
-				w.WriteHeader(http.StatusBadRequest)
-				return
-			}
-			if !isValidTrainAs(*repertoire.TrainAs) {
-				w.WriteHeader(http.StatusBadRequest)
-				return
-			}
-
-			log.Println("creating repertoire for id:", id)
-
-			tx, err := db.Begin()
-			if err != nil {
+				log.Println("database error when creating repertoire:", repertoire, err)
 				w.WriteHeader(http.StatusInternalServerError)
 				return
 			}
-			stmt, err := tx.Prepare("INSERT INTO repertoire(repertoire_id, name, train_as) VALUES (?, ?, ?)")
-			if err != nil {
-				tx.Rollback()
-				w.WriteHeader(http.StatusInternalServerError)
-				return
-			}
-			_, err = stmt.Exec(id, *repertoire.Name, *repertoire.TrainAs)
-			if err != nil {
-				tx.Rollback()
-				w.WriteHeader(http.StatusInternalServerError)
-				return
-			}
-			err = tx.Commit()
-			if err != nil {
-				tx.Rollback()
-				w.WriteHeader(http.StatusInternalServerError)
-				return
-			}
-			stmt.Close()
 
 			w.WriteHeader(http.StatusOK)
+			json.NewEncoder(w).Encode(repertoire)
 		} else {
 			w.WriteHeader(http.StatusMethodNotAllowed)
 		}
